+++ CHANGELOG.md
@@ -6,3 +6,7 @@ Enhancements
 
 * Region select options are now sorted (lukast-akra)
 
+API Changes
+
+* region_options_for_select now takes priority region codes in an options hash
+  under the key 'priority'.